+++ src/protocol/index.ts
@@ -1,3 +1,7 @@
 export { AddressComponentRangeError, MEMORY_REGIONS, MultiSlot, PresetSlot } from "./address";
 export * from "./cc";
+export * from "./delay-clock-rate";
 export * from "./enums";
+export * from "./lfo-clock-rate";
+export * from "./transpose";
+export { Tune } from "./tune";